requirements.txt
<<<<<<< HEAD
flask
python-dotenv
psycopg2-binary
playwright
requestsannotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.76.2
playwright==1.52.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
=======
# flask
# python-dotenv
# psycopg2-binary
# playwright
# anyio==4.9.0
# blinker==1.9.0
# certifi==2025.4.26
# charset-normalizer==3.4.1
# click==8.1.8
# distro==1.9.0
# Flask==3.1.0
# greenlet==3.2.1
# h11==0.16.0
# httpcore==1.0.9
# httpx==0.28.1
# idna==3.10
# itsdangerous==2.2.0
# Jinja2==3.1.6
# jiter==0.9.0
# MarkupSafe==3.0.2
# openai==1.76.2
# playwright==1.52.0
# psycopg2-binary==2.9.10
# pydantic==2.11.4
# pydantic_core==2.33.2
# pyee==13.0.0
# python-dotenv==1.1.0
# requests==2.32.3
# sniffio==1.3.1
# tqdm==4.67.1
# typing-inspection==0.4.0
# typing_extensions==4.13.2
# urllib3==2.4.0
# Werkzeug==3.1.3
# chromadb==0.4.22  # or the version you're using
# numpy<2.0
# sentence-transformers


Flask==3.1.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
playwright==1.52.0
sentence-transformers==4.1.0
chromadb==0.4.22
numpy<2.0
scikit-learn==1.6.1
scipy==1.15.2
transformers==4.51.3
torch==2.7.0
pydantic==2.11.4
requests==2.32.3
openai==1.76.2
uvicorn==0.34.2
fastapi==0.115.12
openpyxl
>>>>>>> 61b9b8f (Patch UI Mapper: crawler, extractor, schema updates, HTML fixes)
